fix: correct integer field types and rework OAuth2/multipart internals
- Change double?/decimal? to int? for integer-only fields across models
  and API methods (ids, page/perPage, intervals, etc.), including the
  AnyOf union variants (TryGetDouble/TryGetDecimal -> TryGetInt)
- Rework OAuth2 authorization-code, client-credentials, and password
  strategies to support both Basic-auth-header and form-body client
  authentication via a new CredentialParamsFactory delegate
- Rewrite multipart form request handling to use explicit
  MultipartParam/BinaryContent models instead of reflection-based file
  detection; add BinaryRequest/BinaryResponse and
  ErrorByteContent/ErrorByteResponse for binary payload support
- Add NonDisposingStream so HttpContent disposal never closes a
  caller-owned stream
- Fix ParameterFlattener to stringify scalar values with invariant
  culture instead of the current culture
- Change client DI registration from AddTransient to AddSingleton in
  ServiceCollectionExtensions
- Regenerate api-reference.md

+++ Api/ApiExports.cs
@@ -107,8 +107,8 @@ public sealed class ApiExports
     /// </para>
     /// </remarks>
     public Task<IReadOnlyList<Invoice>> ListExportedInvoices(string batchId,
-        double? perPage = 100d,
-        double? page = 1d,
+        int? perPage = 100,
+        int? page = 1,
         CancellationToken ct = default) =>
         _rawClient.Execute(_server.Production("/api_exports/invoices/{batch_id}/rows.json"),
             [new TemplateParam("batch_id", batchId)],
@@ -137,8 +137,8 @@ public sealed class ApiExports
     /// </para>
     /// </remarks>
     public Task<IReadOnlyList<ProformaInvoice>> ListExportedProformaInvoices(string batchId,
-        double? perPage = 100d,
-        double? page = 1d,
+        int? perPage = 100,
+        int? page = 1,
         CancellationToken ct = default) =>
         _rawClient.Execute(_server.Production("/api_exports/proforma_invoices/{batch_id}/rows.json"),
             [new TemplateParam("batch_id", batchId)],
@@ -167,8 +167,8 @@ public sealed class ApiExports
     /// </para>
     /// </remarks>
     public Task<IReadOnlyList<Subscription>> ListExportedSubscriptions(string batchId,
-        double? perPage = 100d,
-        double? page = 1d,
+        int? perPage = 100,
+        int? page = 1,
         CancellationToken ct = default) =>
         _rawClient.Execute(_server.Production("/api_exports/subscriptions/{batch_id}/rows.json"),
             [new TemplateParam("batch_id", batchId)],
